+++ mvc/src/main/java/com/interface21/web/DispatcherServletInitializer.java
@@ -1,9 +1,9 @@
-package com.techcourse;
+package com.interface21.web;
 
+import com.interface21.webmvc.servlet.DispatcherServlet;
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import com.interface21.web.WebApplicationInitializer;
 
 /**
  * Base class for {@link WebApplicationInitializer}
